feat(combobox): combobox disclosure button
