+++ packages/components/combobox/src/Combobox.doc.mdx
@@ -40,6 +40,10 @@ import { Combobox } from '@spark-ui/combobox'
       description:
         'The typing area in which the user can type. The input behaviour will differ if `autoComplete` is used or not.',
     },
+    'Combobox.Disclosure': {
+      of: Combobox.Disclosure,
+      description: 'Optional visual button to open and close the combobox items list.',
+    },
     'Combobox.LeadingIcon': {
       of: Combobox.LeadingIcon,
       description: 'Prepend a decorative icon inside the input (to the left).',
